chore(deps): move solver to miden 0.15.2 (client tag) + protocol 0.15.3
Pin to the released miden-client v0.15.2 git tag (+ sqlite-store) and
miden-protocol/standards/testing 0.15.3 from crates.io (the versions the
v0.15.2 client manifest declares), with miden-core 0.23 / miden-crypto
0.25. v0.15.2 is the client release that fixes the v0.15.1<->0.15.3
TransportNote.after_block_num mismatch, so it finally builds.

Production miden-protocol/miden-client deps carry NO `testing` feature;
tests re-enable it via [dev-dependencies] (resolver = 2 keeps it out of
the prod build).

Test-code API migration for 0.15.3:
- Felt::new is now fallible (consume-script).
- AssetAmount is a newtype: sum via u64::from(asset.amount()).
- AccountStorageMode -> AccountType (Public/Private) for the test wallet/
  faucet helpers.
- NoteAttachment::default() -> None (build_pswap_create now takes
  Option<NoteAttachment>).
- NoteId::new now needs commitment+metadata; test ids via try_from_hex.
- testing const ACCOUNT_ID_NETWORK_FUNGIBLE_FAUCET removed -> PRIVATE.

Verified: cargo build --bin solver-bin (prod) clean; cargo build
--workspace --all-targets clean (incl. integration tests); 163 lib tests
pass. Integration tests COMPILE but were not run here (need a release
build for proof-gen; blocked by low disk this session) — run
`cargo test --release` to confirm end-to-end.

+++ crates/consume-script/tests/consume_asset_test.rs
@@ -139,7 +139,7 @@ async fn consume_asset_script_captures_surplus() -> anyhow::Result<()> {
     let usdc_added: u64 = added
         .iter()
         .filter_map(|a| match a {
-            Asset::Fungible(f) if f.faucet_id() == usdc_faucet.id() => Some(f.amount()),
+            Asset::Fungible(f) if f.faucet_id() == usdc_faucet.id() => Some(u64::from(f.amount())),
             _ => None,
         })
         .sum();
